Added controllers for all routes to modularize codebase
