+++ .gitignore
@@ -41,3 +41,4 @@ testem.log
 #System Files
 .DS_Store
 Thumbs.db
+controllers/mainRoutesController.js___jb_old___
